Use nvcomp conda package. (#245)
This PR uses conda-forge packages of `nvcomp` rather than fetching a tarball. This means that the nvcomp binary should not be shipped in the kvikio conda package, but is instead listed as a dependency. This will reduce kvikio's conda package size.

Resolves #242.

Authors:
  - Bradley Dice (https://github.com/bdice)

Approvers:
  - https://github.com/jakirkham
  - Ray Douglass (https://github.com/raydouglass)

URL: https://github.com/rapidsai/kvikio/pull/245