Support zarr-python 3.x (#646)
This starts the implementation of a zarr-python 3.x compatible GDSStore. Not ready for review yet, lots of things to address first.

We have a few things to figure out:

1. Should kvikio try to support zarr-python 2.x and zarr-python 3.x?
2. If we try to support both, do we have separate `kvikio.zarr` and `kvikio.zarr_v3` namespaces?

This should be doable: We can have a `kvikio.zarr` check the version of zarr-python. For 2.x we'd import the old implementation, for 3.x we'd implement the new implementation.

There are some complications to this, though. It makes the classes / functions and their signatures you get from `kvikio.zarr` dynamic rather than statically known. Also makes testing this harder.

And the `zarr_v3` name is a misnomer, since zarr-python supports both zarr v3 and v2.

3. It seems like the replacement for `NVCompCompressor` might end up in zarr-python, rather than kvikio.
4. Are we OK with `GDSStore` just supporting local files? Would we want to try to support remote files as well? (IMO, defer that for now)

Closes https://github.com/rapidsai/kvikio/issues/582.

Authors:
  - Tom Augspurger (https://github.com/TomAugspurger)

Approvers:
  - Mads R. B. Kristensen (https://github.com/madsbk)
  - Bradley Dice (https://github.com/bdice)

URL: https://github.com/rapidsai/kvikio/pull/646