Fix benchmark code on MIG (#703)
In UCX-Py and UCXX, we skip querying memory info on MIG devices as the approach we use on other hardware doesn't work for MIG. However KvikIO did not skip querying memory info on MIG, which causes this code to fail there.

To fix this issue, we adapt the approach used in UCX-Py and UCXX to KvikIO, which allows this code to run on MIG.

Additionally the code used one `try...except...` to handle everything from `import`ing to querying the device. With the MIG check too, this becomes kind of unwieldy for one `try...except...`. So this moves the `import` to the top-level and breaks up the `try...except...`. This makes the logic of the code and the cases it handles a bit more legible to readers 

Fixes https://github.com/rapidsai/kvikio/issues/701

Authors:
  - https://github.com/jakirkham

Approvers:
  - Mads R. B. Kristensen (https://github.com/madsbk)

URL: https://github.com/rapidsai/kvikio/pull/703