refactor(wheels): vendor nvcomp as temporary workaround (#702)
xref rapidsai/build-planning#120

RAPIDS is blocked on Python 3.13 support because there is no `nvcomp` release that (yet) supports Python 3.13.

This is a temporary (and wacky) workaround that re-vendors `nvcomp` into
`libkvikio` and removes the runtime dependency on the `nvcomp` wheel, removing
with it the block on Python 3.13.

Once there's an updated `nvcomp` release, we can undo this.

Notes:

I am no kind of CMake expert, and I'm sure this can be done better.

Using `kvikio-exports` as the `BUILD_EXPORT_SET` is no good because it installs
3+ copies of the same 100mb shared object and balloons the size of the wheel.
There are _two_ copies of `libnvcomp.so` as it stands, but they compress down
well enough.  If anyone has any pointers on removing the remaining `so`, I'm
all ears.  I know it's possible because the `nvcomp` wheel only has a single
`libnvcomp.so.4`, but also, this is a temporary patch and it's good enough.

I've added in the `WheelHelpers.cmake` module because it does what I want, but
if there's something simpler, we can try that, too.

Authors:
  - Gil Forsyth (https://github.com/gforsyth)

Approvers:
  - Bradley Dice (https://github.com/bdice)
  - Vyas Ramasubramani (https://github.com/vyasr)

URL: https://github.com/rapidsai/kvikio/pull/702