Add support for temporary AWS credentials to access S3 (#693)
This PR adds the `x-amz-security-token` header which is needed when using temporary credentials (These start with `ASIA`). According to the [curl docs](https://curl.se/libcurl/c/CURLOPT_HTTPHEADER.html), the list of headers needs to be freed after use, So I added a destructor to `S3Endpoint` that does that.
Important to note is that I added a parameter for the session token, but it comes directly after the other credential parameters and _before_ the endpoint override, making this a breaking change for code that interacts directly with these C++ APIs! But it makes it consistent with the other values; otherwise 1 of them would need to be passed only via the env var.
It seems there might be another API change here https://github.com/rapidsai/kvikio/pull/700 so maybe both changes can be done in the same release.
I hope this is useful for others too, looking forward to any comments!

Fixes #584 (@TomAugspurger)

Authors:
  - Joost Hoozemans (https://github.com/joosthooz)

Approvers:
  - Tianyu Liu (https://github.com/kingcrimsontianyu)
  - Bradley Dice (https://github.com/bdice)

URL: https://github.com/rapidsai/kvikio/pull/693