Revert "Provide API to detect C2C platform and add NVML shim" (#731)
Reverts rapidsai/kvikio#721

We're seeing some downstream build breakages in cudf 25.08 runs. It looks like this PR is requiring that NVML be installed in the build environment, which was not previously the case, and since cudf is now frozen for 25.06 we don't want to have to modify its build setup right now. We can reopen this PR for 25.08 and then get the issues fixed there.

Authors:
  - Vyas Ramasubramani (https://github.com/vyasr)

Approvers:
  - Tianyu Liu (https://github.com/kingcrimsontianyu)
  - Bradley Dice (https://github.com/bdice)

URL: https://github.com/rapidsai/kvikio/pull/731