Drop unneeded version check for batch/stream APIs
As of CUDA 12.2.0 all of these functions are available. So drop the
version checks around them.

Though still get the version (if possible) as it is defined here and
used throughout the codebase. If it is not possible to get the version,
then this is CUDA 12.2 (as CUDA 12.3 defines the function). So set the
version to cuFile 1.7.0, which was included in CUDA 12.2.